Refactor: convert order items into a table from array
- Order items being an array was a pain since we can't really use joins to fetch the item directly
- We had to do 2 queries, one for fetching all items in order and then fetching all items and then pairing them up manually in frontend, which is retarded.
- this refactor hopefully helps mitigate that, while still being efficient enough

+++ src/enums/common.rs
@@ -5,7 +5,7 @@ use crate::models::common::OrderItems;
 #[derive(Serialize, ToSchema)]
 pub struct ActiveItemCount {
     pub item_id: i32,
-    pub num_ordered: i32,
+    pub num_ordered: i64,
 }
 
 #[derive(Serialize, ToSchema)]
